Move navbar outside of div.container in all places
The `div.container` for containing the main content width is applied
inconsistently. Sometimes it includes the navbar and sometimes it
doesn't. This is partly from the conversion of some pages to jsx. It's
also difficult to notice this because the navbar has fixed positioning
that breaks it out of document flow and allows it to have full 100%
viewport width regardless of its container width.