Clean up flexbox settings for Values and Sourcing
Set a flex-basis (min-width in this case) on the `.left` and `.right`
content blocks, and adjusted flow to make them wrap. This may alter the
width and positioning of the images and text slightly, but is more
robust and requires fewer special rules for individual elements (which
were removed).

These changes also affect the `.block` section on the teas-products
page, which needs some additional cleanup for mobile now that the
content stacks.